updated iperf3 reporting to include retransmits and cpu usage
  * iperf3 tests now report on sender retransmits as well as sender
cpu and receiver cpu usage during tests.  I adopted the format already
in use for qperf tests for reporting multiple data points, grouping
them in parenthesis and separating them with semi-colons.  In the future
it will probably make sense to update the output format to make it
more friendly for graphing utilities.